requirements.txt
asgiref==3.8.1
Django==5.0.6
django-bootstrap5==24.2
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg==3.1.19
psycopg-binary==3.1.19
pycountry==23.12.11
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
